refactor: move old functions back to where they came from
